Optimize process.one_iteration cache handling
This optimization is expected to improve performance by ~50%. The inefficiency
is that for every WSMuon, the entire event cache is examined, even though the
vast majority of events in the cache have already been assigned a
dt_next_WSMuon. They are in the cache because they're waiting for a
next_DelayedLike. So this optimization tracks how many events in the cache
already have dt_next_WSMuon, then skips ahead to the first "new" cached event
when assigning dt_next_WSMuon. (This is achieved using deque.rotate on the
event_cache object, and then breaking out of the loop when the first "old"
event is encountered.)

Using line_profiler, I determined that the WSMuon cache loop was taking
75% of the execution time of the one_iteration function. In the new
optimized version, the WSMuon cache loop takes only 5% of the execution time.
This reduced the time to execute the full event selection on 10000 events from
~30 seconds to ~10 seconds.