Amend previous commit: Store up to 100 recent prompt-likes
Turns out that it's not that uncommon in a given file for up to 50 or 60
prompt-like events to cluster within 400us. And possibly even more. So I'm
compromising, by extending the array up to 100 entries, and then protecting it
from overflows by limiting the deque that recent prompt-likes are stored in to
100 events. (Newer entries push out older entries.)