Update prediction to account for differences in livetime, LBNL-style
The intuitive way to account for livetime differences is to change
the reactor flux for each AD.
However the LBNL fitter assumes the same flux for each AD
and then multiplies the final predicted rate by a given AD's livetime.
So I am doing the same, except instead of tracking the rate all the way
through, I use total events and then correct by the ratio of far/near
livetime.